Do not instantiate the synchronizer each time.
This has the nice effect of letting the persistent-connection
reuse the existing connection, avoiding the ssl handshake overhead
each time we try to synchronize.

This can be traced by logging the instantiation of HttpClientBase
in u1db.remote

I *think* we should be fine with the timeouts as long as we keep
the sync period along the 1 min we are doing now. For other cases,
we should look into how to override the default timeout in httplib
(used by u1db http_client).