[bug] do not block when getting doc for async enc
Previous to this change, the actual encryption method used to run on its own
thread. When the close method was called from another thread, the queue could
be deleted after the encryption method loop had started, but before the queue
was checked for new items.

By removing that thread and moving the encryption loop to the reactor, that
race condition should disappear.

Closes: #7088.