[bug] catch 'point is undefined' error on couch
When trying to use an unexisting list function, to alter a view, the error is
not a missing document error, but an obscure "TypeError" saying that "point is
undefined" because of the way the javascript code in couchdb server tries to
find the list function.

This commit adds a catch for that error and raises the proper exception in the
soledad couch module.