[feat] update profile-sync script, add plop profiling
Updating the profile-sync script to the latest deferred-based sync.

- Added a couple of options to have finer control about what output to
  get.
- Add support for the plop profiler https://pypi.python.org/pypi/plop/
- To get meaningful plop profiles, make sure to disable the system stats
  collection, like this::

  ./profile-sync.py --no-stats --plop -b /tmp/syncdata/ -p sikret user@provider

A good practice for this is having a pre-seeded soledad account
(currently you have to do that through the wizard, a cli will be very
nice to have in the coming future) with a known amount of data (for
instance, sending some mails with known payload weight).

It is VERY IMPORTANT that you *NEVER* process the data in this account
(ie, do not ever log in with a mail client, for instance, since that
will alter the original documents).

In order to have comparable results, you should always run this sync
script in similar conditions. Ideally, on a machine with runlevel 1.

Also, make sure of deleting the contents in the folder if you are
not using a temporal dir.