[refactor] remove user_id argument from Soledad init
The constructor method of Soledad was receiving two arguments for user
id. One of them was optional with None as default. It could cause an
inconsistent state with uuid set but userid unset.

This change remove the optional user_id argument from initialization
method and return the uuid if anyone call Soledad.userid method.