[pkg] remove pixelated from requirements-latest.pip
modifying original PR [0] by cristoph to account for the recent
vendoring of l2db code, which means we no longer depend on u1db/dirspec.

I expect the whole mess about the venv setup to be further simplified
pretty soon, since we are going to merge most of the leap.* packages
into a couple of repos.

[0] https://github.com/leapcode/soledad/pull/327