[feat] do not use couch views for sync metadata
When compared to plain couch document get, the use of the simplest view
functions takes around double the time, while the use of the simplest
list function can take more than 8 times:

  get 100 docs:
    total: 0.440337 secs
    mean:  0.004403
  query 100 views:
    total: 0.911425 secs
    mean:  0.009114
  query 100 lists:
    total: 3.711537 secs
    mean:  0.037115

Besides that, the current implementation of sync metadata storage over
couch is dependent of timestamps of document puts, what can lead to
metadata corruption if the clock of the system is changed for any
reason.

Because of these reasons, we seek to change the implementation of
database metadata. This commit implements the storage of transaction log
data on couch documents with special ids, in the form "gen-xxxxxxxxxx",
where the x's are replaced by the generation index.

Each generation document holds a dictionary containing the generation,
doc_id and transaction_id for the changed document. For each modified
document, a generation document is inserted holding the transaction
metadata.