[bug] fix order of multipart serialization when writing to couch
The couch backend makes use of attachments and multipart structure for
writing the document to the couch database. For that to work, the order
in which attachments are described must match the actual order in which
attachments are written to the couch http stream.

This was not being properly taken care of, and eventually the json
serializer was arbitrarilly ordering the attachments description in a
way that it didn't match the actual order of attachments writing.

This commit fixes that by using json.dumps() sort_keys parameter and
making sure conflicts are always written before content.