[pkg] add soledad to the ssl-cert group
Now that uid an gid of the application are changed earlier by systemd
(instead of twistd), soledad will fail to start if it doesn't have
permission to read the ssl certificate with the new uid/gid.

This commit adds soledad to the ssl-cert group, which is the group of
the certificate as set by the platform.