[bug] effectively load the configuration for the app
the code for passing the configuration to the couch initialization was
never called. it seems the entrypoint module wasn't finally hooked as
expected. I think this fixes the problem, but further review is needed
here: either the entrypoint module is to be used, or it better is
removed. in the first case, this workaround probably needs to be
reverted.