[bug] reuse wsgi threadpool
it seems evident that the functions were thought to pass a threadpool
along, but it finally wasn't properly passed and so there was a new
threadpool created to handle every resource.

I have removed the creation from the factory because I don't think it
makes sense to create a threadpool on the fly, it's prone to errors.

- Resolves: #8774