[pkg] do not pass weird information to install_requires
this commit deliberately *duplicates* the dependency information between
the setup.py and requirements.pip.

The rationale behind it is that one of them is the generic dependency
information, to be passed to system helpers in order to package it.

The requirements.pip has any pinning information that we want to impose
during development, and it behaves as a well-known set.

Read https://caremad.io/posts/2013/07/setup-vs-requirement/ for more
explanations.

- Resolves: #8830