[bug] incoming doesnt accept encryption schemes
According to /incoming specification, this parameter doesn't exists. It
was added on MX side and brought here on transition. This commit
explicitly removes it, leaving a track on history that it got removed
during transition. Further work on MX will also remove the "NONE"
encryption scheme, which is part of dead code.