[pkg] add a symlink to soledad-create-userdb for backwards compatibility
We have changed the name of the script, to qualify it with the
'soledad' preffix.

Here I am creating a symlink with the name of the previous script, for
backwards-compatibility reasons - basically to avoid breaking unattended
upgrades.

(Note that the path is hardcoded in the config file deployed by the
platform, so it needs to be changed in the platform accordingly too).

This symlink should be deprecated after a couple of release cycles.