[pkg] Add ssl-cert to soledad-server deb deps
The soledad user needs to read the ssl certificate
private key stored in `/etc/x509/keys/leap.key`,
which is accessible by users in the `ssl-cert` group,
added by the `ssl-cert` package.
On installation, the soledad user is added to that group
so the ssl-package needs to be installed by then.

- Resolves: #8849