[benchmarks] limit number of runs of sqlcipher benchmark tests
As sqlcipher benchmark tests take longer, we want to limit the number of
repetitions. Previous to this change, these tests were being calibrated
automatically and would run 5 times becuase it is the default minimum
number of times for pytest-benchmark. By changing the runner to pedantic
mode, now they will be run 4 times, the same number of times as
benchmark sync tests.