[test] add --subdir option and blacklist
Use subdir path to select/deselect benchmark tests.

Because of the way pytest and pytest benchamrk select/deselect tests, it
makes more sense for this test suite to use a --subdir option to select
only a subset of tests and a blacklist of subdirectories that should not
be run unless explicitelly asked to.

This commit adds a --subdir option, that will deselect all tests no in
./tests/<subdir> when the option is passed to pytest. Also, a blacklist
is added so, unless explicitelly added as a --subdir, tests in
blacklisted subdirectories will not be run.

The reasons for this modification are:

- pytest-benchmarks selects/deselects tests based on the type of their
  fixtures. This means that fixtures have to be instantiated previous to
  test selection/deselection, something that may cause side effects that
  we want to avoid.

- the usual test suite is run for all benchmarks, while some tests
  should only be run in specific situations (as is the case of
  benchmarks and responsiveness tests). It is saner to implement subdir
  selection than marking all tests with pytest marks and using them to
  select/deselect.