[test] selectivelly start events server on benchmarks
Soledad's events server might be a pain for tests because if it is not
properly initialized and shut down there might be parts of the server
missing or leftover in the reactor after tests are run. With the recent
changes to the way tests are invocated, there's the need to further
isolate events server startup for benchmarks, otherwise it will conflict
with other tests. This commit moves the events server setup to inside
the pytest_collection_modifyitems() hook, where we can check if the
benchmark subdir was selected and properly start the events server if
needed.