[pkg] make debian package report the correct version
The Soledad python source currently makes use of versioneer to calculate
versions dynamically and hardcode versions upon creation of the python
package. Unfortunatlly, this procedure does not behave well with debian
packaging, and we need to manually generate/replace the _version.py file
so the code that ends up in the debian package reports the correct
version of the software.

This commit adds a script to replace the _version.py file with a static
one and rules during packaging to run that script during preparation
stage of debian helper.