[test] TAC test now checks for server (with retries)
After we fixed the server to wait for checks before listening on ports
(201ef7a9b979f8c8efaedbe542c631944d8956f4), the TAC test started failing
randomly in our CI, probably because sometimes the server takes more
than 1 second to startup in the CI setup (docker in a vm).

This commit adds a check with retry, and so the test will retry 10 times
with an interval of 1 second before failing.