[test] always use a tempdir for server fixture config file
When starting up a Soledad Server through the use of a fixture in tests,
a config file is created with the options needed for that test. As the
config file was being created in /etc and not being removed, this was
influencing servers started by other tests that don't use that fixture.

By always using a temporary directory in the server fixture, we make
sure that the config file is used only for that test and is removed
after the test is finished.

Closes: #8995